fix(hoisting): allow hoisted functions to be used before being defined
`no-use-before-define` is great to prevent confusing use of hoisted variables, but it's pretty common to but the meat of a a file at the top and then define helper functions at the bottom out of the way. Moreover, circular-ish dependencies caused by event-based programming often make it difficult to completely avoid using a function before it's defined. I'm ambivalent on classes, but they seem more like functions that variables.